requirements.txt
Django==4.1.10
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
python-decouple==3.8
django-filter==23.3
Pillow==9.5.0
django-environ
mssql-django